pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brainfile"
version = "0.2.0"
description = "Python library for the Brainfile task management protocol"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Brainfile Contributors" }
]
keywords = ["brainfile", "task-management", "kanban", "markdown", "yaml", "ai-agents", "mcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "ruamel.yaml>=0.18",
]

[project.optional-dependencies]
watch = [
    "watchdog>=3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "mypy",
    "ruff",
    "watchdog>=3.0",
]

[project.urls]
Homepage = "https://brainfile.md"
Repository = "https://github.com/brainfile/py"
Issues = "https://github.com/brainfile/py/issues"
Documentation = "https://github.com/brainfile/protocol"

[tool.hatch.build.targets.wheel]
packages = ["src/brainfile"]

[tool.mypy]
strict = true
python_version = "3.11"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=brainfile --cov-report=term-missing"
